[I-10] Scheduled deletion plus small fixes

+++ src/main/java/com/example/orderdemo/service/impl/UserServiceImpl.java
@@ -5,9 +5,11 @@ import com.example.orderdemo.model.dto.UserDto;
 import com.example.orderdemo.repository.UserRepository;
 import com.example.orderdemo.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.apache.velocity.exception.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -18,6 +20,8 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
 
+    private static final String MESSAGE = "Entity is not presented.";
+
     @Override
     public void createUser(UserDto user) {
         userRepository.save(userMapper.mapToUser(user));
@@ -25,6 +29,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<UserDto> getUserList() {
+        if (userRepository.findAll().isEmpty()) {
+            throw new ResourceNotFoundException(MESSAGE);
+        }
         return userRepository.findAll()
                 .stream()
                 .map(userMapper::mapToUserDto)
@@ -33,11 +40,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto getUserById(UUID id) {
+        if (Objects.equals(userRepository.findById(id), null)) {
+            throw new ResourceNotFoundException(MESSAGE);
+        }
         return userMapper.mapToUserDto(userRepository.findById(id));
     }
 
     @Override
     public void deleteUser(UUID id) {
+        if (Objects.equals(userRepository.findById(id), null)) {
+            throw new ResourceNotFoundException(MESSAGE);
+        }
         userRepository.deleteById(id);
     }
 }
